+++ leet/streak.go
@@ -1,4 +1,4 @@
-package game
+package leet
 
 import (
 	"database/sql"
@@ -131,9 +131,9 @@ func GetActiveStreaks(db *sql.DB) ([]Streak, error) {
 	timeNow := time.Now()
 	var timeString string
 	if timeNow.Hour() < 13 || (timeNow.Hour() == 13 && timeNow.Minute() < 38) {
-	    timeString = time.Now().AddDate(0, 0, -1).Format("2006-01-02")
+		timeString = time.Now().AddDate(0, 0, -1).Format("2006-01-02")
 	} else {
-	    timeString = time.Now().Format("2006-01-02")
+		timeString = time.Now().Format("2006-01-02")
 	}
 
 	rows, err := db.Query(`
